+++ CHANGELOG.md
@@ -2,4 +2,11 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
-## Unreleased
+## [1.0.0] - 2017-06-07
+
+### Added
+
+- Functionality defined on `StandardObjectInterface`.
+- An object proxy trait and standard object implementation is provided, that is intended to wrap a `stdClass` object.
+- Utilities provided on static methods on the `Obj` class.
+- Ability to `json_decode` to a standard object using the `Obj::decode()` method.